cast to float32

+++ baselines/baselines/common/input.py
@@ -54,7 +54,8 @@ def encode_observation(ob_space, placeholder):
     if isinstance(ob_space, Discrete):
         return tf.to_float(tf.one_hot(placeholder, ob_space.n))
     elif isinstance(ob_space, Box):
-        return tf.to_float(placeholder)
+        #return tf.to_float(placeholder)
+        return tf.cast(placeholder, tf.float32)
     elif isinstance(ob_space, MultiDiscrete):
         placeholder = tf.cast(placeholder, tf.int32)
         one_hots = [tf.to_float(tf.one_hot(placeholder[..., i], ob_space.nvec[i])) for i in range(placeholder.shape[-1])]
